Implement Yoga::InputText & Yoga::InputTextField, change Parser float type do double
As of now a quite simple wrapper of ImGui::TextInputEx into our Yoga::Item system.
Validators are also implemented, they do a simple math equation processing, meaning user
can use +,-,*,/ symbols while editing a value.
In the backend ImGui works with std::string.

Yoga::InputText only provides bare text input with validators, it is then composed
into Yoga::InputTextField which also decorates the component.

Unit tests for Validator classes are also added

+++ src/slic3r-domain/include/Slic3r/Domain/Expr/ExprAst.hpp
@@ -21,7 +21,7 @@ struct VarRef;
 
 using ExprAst = boost::variant<
     bool,
-    float,
+    double,
     std::string,
     RegEx,
     boost::recursive_wrapper<Binary>,
