+++ src/slic3r-shared/src/Slic3r/Biz/Preset/ValueMapBuilder.hpp
@@ -13,8 +13,8 @@ void append_value(Expr::ValueMap& values, const std::string& name, const std::va
     Expr::Value val;
     if (std::holds_alternative<std::string>(v)) {
         val = std::get<std::string>(v);
-    } else if (std::holds_alternative<float>(v)) {
-        val = std::get<float>(v);
+    } else if (std::holds_alternative<double>(v)) {
+        val = std::get<double>(v);
     } else if (std::holds_alternative<bool>(v)) {
         val = std::get<bool>(v);
     } else {
